fix formatting

+++ DWN/main.py
@@ -168,8 +168,9 @@ def evaluate_model(x_train, y_train, X_test, y_test, encoder, start_time):
             optimizer.zero_grad()
 
             indices = permutation[i : i + BATCH_SIZE]
-            batch_x, batch_y = x_train[indices].cuda(device), y_train[indices].cuda(
-                device
+            batch_x, batch_y = (
+                x_train[indices].cuda(device).float(),
+                y_train[indices].cuda(device).float(),
             )
             outputs = model(batch_x)
             loss = cross_entropy(outputs, batch_y)
